test - removing dynamic time
